requirements.txt
asgiref==3.7.2
boto3==1.33.5
botocore==1.33.5
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==4.4.2
Django==4.2.7
django-appconf==1.0.5
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-environ==0.11.2
django-fastdev==1.8.0
django-js-asset==2.1.0
django-phonenumber-field==7.2.0
django-storages==1.14.2
djangorestframework==3.14.0
humanize==4.8.0
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
jmespath==1.0.1
moviepy==1.0.3
mysqlclient==2.2.0
numpy==1.26.2
pathvalidate==2.3.0
phonenumbers==8.13.25
Pillow==10.0.1
proglog==0.1.10
psutil==5.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
